// admin-service/src/dto/user.dto.ts
export class UserDto{
    name: string;
    email: string;
    OTP: string;
    password: string;  
    oldpassword : string;
    newpassword : string;
    user_id: string;
    countryCode:string;
    createdAt: Date;
    updatedAt: Date;
    profileImg: string;
    contact: string;
    status:string;
    pageNo: string;
    perPage: string;
    search: string;
    toDate: string;
    fromDate: string;
    language:string;
    type: string;
    
}